Moved dashboard state components to state folder
