+++ public/app/features/alerting/StateHistory.tsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import alertDef from './state/alertDef';
 import { getBackendSrv } from 'app/core/services/backend_srv';
-import { DashboardModel } from '../dashboard/dashboard_model';
+import { DashboardModel } from '../dashboard/state/DashboardModel';
 import appEvents from '../../core/app_events';
 
 interface Props {
