+++ public/app/features/dashboard/state/DashboardModel.test.ts
@@ -1,6 +1,6 @@
 import _ from 'lodash';
-import { DashboardModel } from '../dashboard_model';
-import { PanelModel } from '../panel_model';
+import { DashboardModel } from '../state/DashboardModel';
+import { PanelModel } from '../state/PanelModel';
 
 jest.mock('app/core/services/context_srv', () => ({}));
