+++ public/app/features/dashboard/state/DashboardModel.ts
@@ -7,8 +7,8 @@ import { Emitter } from 'app/core/utils/emitter';
 import { contextSrv } from 'app/core/services/context_srv';
 import sortByKeys from 'app/core/utils/sort_by_keys';
 
-import { PanelModel } from './panel_model';
-import { DashboardMigrator } from './dashboard_migration';
+import { PanelModel } from './PanelModel';
+import { DashboardMigrator } from './DashboardMigrator';
 import { TimeRange } from '@grafana/ui/src';
 
 export class DashboardModel {
