+++ public/app/features/dashboard/state/PanelModel.ts
