+++ public/app/features/dashboard/state/index.ts
@@ -1,0 +1,2 @@
+export { DashboardModel } from './DashboardModel';
+export { PanelModel } from './PanelModel';
