+++ public/test/specs/helpers.ts
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import config from 'app/core/config';
 import * as dateMath from 'app/core/utils/datemath';
 import { angularMocks, sinon } from '../lib/common';
-import { PanelModel } from 'app/features/dashboard/panel_model';
+import { PanelModel } from 'app/features/dashboard/state/PanelModel';
 
 export function ControllerTestContext(this: any) {
   const self = this;
